+++ client/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
-import store from './store.js';
+import store from './store';
 
 // Material UI
 import darkBaseTheme from 'material-ui/styles/baseThemes/darkBaseTheme';
@@ -9,7 +9,7 @@ import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 import getMuiTheme from 'material-ui/styles/getMuiTheme';
 
 // Components
-import App from './components/app.js';
+import App from './components/app';
 
 const ThemedIndex = () => (
   <MuiThemeProvider muiTheme={getMuiTheme(darkBaseTheme)}>
